+++ src/pagoda/geometry/operations/clip_geometry.cpp
@@ -2,8 +2,6 @@
 
 #include <pagoda/geometry/geometry_component.h>
 #include <pagoda/geometry/geometry_system.h>
-#include <pagoda/objects/hierarchical_component.h>
-#include <pagoda/objects/hierarchical_system.h>
 #include <pagoda/objects/procedural_object_system.h>
 
 #include <pagoda/dynamic/float_value.h>
@@ -41,7 +39,6 @@ void ClipGeometry::DoWork()
 	START_PROFILE;
 
 	auto geometrySystem = m_proceduralObjectSystem->GetComponentSystem<GeometrySystem>();
-	auto hierarchicalSystem = m_proceduralObjectSystem->GetComponentSystem<HierarchicalSystem>();
 
 	while (HasInput(inputGeometry))
 	{
@@ -54,12 +51,12 @@ void ClipGeometry::DoWork()
 		auto inGeometryComponent = geometrySystem->GetComponentAs<GeometryComponent>(inObject);
 		GeometryPtr inGeometry = inGeometryComponent->GetGeometry();
 
-		auto frontProceduralObject = CreateOutputProceduralObject(frontGeometry);
+		auto frontProceduralObject = CreateOutputProceduralObject(inObject, frontGeometry);
 		auto frontGeometryComponent = geometrySystem->CreateComponentAs<GeometryComponent>(frontProceduralObject);
 		auto front = std::make_shared<Geometry>();
 		frontGeometryComponent->SetGeometry(front);
 
-		auto backProceduralObject = CreateOutputProceduralObject(backGeometry);
+		auto backProceduralObject = CreateOutputProceduralObject(inObject, backGeometry);
 		auto backGeometryComponent = geometrySystem->CreateComponentAs<GeometryComponent>(backProceduralObject);
 		auto back = std::make_shared<Geometry>();
 		backGeometryComponent->SetGeometry(back);
@@ -70,13 +67,6 @@ void ClipGeometry::DoWork()
 		    Scope::FromGeometryAndConstrainedRotation(front, inGeometryComponent->GetScope().GetRotation()));
 		backGeometryComponent->SetScope(
 		    Scope::FromGeometryAndConstrainedRotation(back, inGeometryComponent->GetScope().GetRotation()));
-
-		auto parentHierarchicalComponent = hierarchicalSystem->GetComponentAs<HierarchicalComponent>(inObject);
-		for (const auto& object : {frontProceduralObject, backProceduralObject})
-		{
-			auto hierarchicalComponent = hierarchicalSystem->CreateComponentAs<HierarchicalComponent>(object);
-			hierarchicalSystem->SetParent(hierarchicalComponent, parentHierarchicalComponent);
-		}
 	}
 }
 }  // namespace pagoda::geometry::operations