+++ src/pagoda/geometry/operations/extrude_geometry.cpp
@@ -4,8 +4,6 @@
 #include <pagoda/geometry/geometry_component.h>
 #include <pagoda/geometry/geometry_system.h>
 
-#include <pagoda/objects/hierarchical_component.h>
-#include <pagoda/objects/hierarchical_system.h>
 #include <pagoda/objects/procedural_component.h>
 #include <pagoda/objects/procedural_object_system.h>
 
@@ -41,7 +39,6 @@ void ExtrudeGeometry::DoWork()
 {
 	START_PROFILE;
 	auto geometrySystem = m_proceduralObjectSystem->GetComponentSystem<GeometrySystem>();
-	auto hierarchicalSystem = m_proceduralObjectSystem->GetComponentSystem<HierarchicalSystem>();
 
 	while (HasInput(input_geometry))
 	{
@@ -52,7 +49,7 @@ void ExtrudeGeometry::DoWork()
 		Extrusion<Geometry> extrude(extrusion_amount);
 
 		// Geometry
-		ProceduralObjectPtr out_object = CreateOutputProceduralObject(output_geometry);
+		ProceduralObjectPtr out_object = CreateOutputProceduralObject(in_object, output_geometry);
 		std::shared_ptr<GeometryComponent> geometry_component =
 		    geometrySystem->CreateComponentAs<GeometryComponent>(out_object);
 		std::shared_ptr<GeometryComponent> in_geometry_component =
@@ -65,14 +62,6 @@ void ExtrudeGeometry::DoWork()
 		geometry_component->SetGeometry(out_geometry);
 		geometry_component->SetScope(
 		    Scope::FromGeometryAndConstrainedRotation(out_geometry, in_geometry_component->GetScope().GetRotation()));
-
-		// Hierarchy
-		std::shared_ptr<HierarchicalComponent> in_hierarchical_component =
-		    hierarchicalSystem->GetComponentAs<HierarchicalComponent>(in_object);
-		std::shared_ptr<HierarchicalComponent> out_hierarchical_component =
-		    hierarchicalSystem->CreateComponentAs<HierarchicalComponent>(out_object);
-
-		hierarchicalSystem->SetParent(out_hierarchical_component, in_hierarchical_component);
 	}
 }
 }  // namespace pagoda::geometry::operations