+++ src/pagoda/geometry/operations/scope_texture_projection.cpp
@@ -4,8 +4,6 @@
 #include <pagoda/geometry/geometry_component.h>
 #include <pagoda/geometry/geometry_system.h>
 
-#include <pagoda/objects/hierarchical_component.h>
-#include <pagoda/objects/hierarchical_system.h>
 #include <pagoda/objects/procedural_object_system.h>
 
 #include <pagoda/dynamic/boolean_value.h>
@@ -52,12 +50,11 @@ void ScopeTextureProjection::DoWork()
 	START_PROFILE;
 
 	auto geometrySystem = m_proceduralObjectSystem->GetComponentSystem<GeometrySystem>();
-	auto hierarchicalSystem = m_proceduralObjectSystem->GetComponentSystem<HierarchicalSystem>();
 
 	while (HasInput(s_inputGeometry))
 	{
 		ProceduralObjectPtr inObject = GetInputProceduralObject(s_inputGeometry);
-		ProceduralObjectPtr outObject = CreateOutputProceduralObject(s_outputGeometry);
+		ProceduralObjectPtr outObject = CreateOutputProceduralObject(inObject, s_outputGeometry);
 
 		auto inGeometryComponent = geometrySystem->GetComponentAs<GeometryComponent>(inObject);
 		GeometryPtr inGeometry = inGeometryComponent->GetGeometry();
@@ -115,10 +112,6 @@ void ScopeTextureProjection::DoWork()
 		{
 			outGeometry->GetVertexAttributes(*i).m_texCoords = ptp.GetProjection(outGeometry->GetPosition(*i));
 		}
-
-		auto inHierarchicalComponent = hierarchicalSystem->GetComponentAs<HierarchicalComponent>(inObject);
-		auto outHierarchicalComponent = hierarchicalSystem->CreateComponentAs<HierarchicalComponent>(outObject);
-		hierarchicalSystem->SetParent(inHierarchicalComponent, outHierarchicalComponent);
 	}
 }
 }  // namespace pagoda::geometry::operations