+++ src/pagoda/material/material.cpp
@@ -13,9 +13,19 @@ namespace pagoda::material
 {
 Material::Material() {}
 
-void Material::SetTexture(const std::string &textureFile) { m_texture = std::make_shared<Image>(textureFile); }
+void Material::SetTexture(int slot, const std::string &textureFile)
+{
+	m_textures[slot] = std::make_shared<Image>(textureFile);
+}
 
-void Material::SetTexture(const image::ImagePtr texture) { m_texture = texture; }
-
-ImagePtr Material::GetTexture() const { return m_texture; }
+void Material::SetTexture(int slot, const image::ImagePtr texture) { m_textures[slot] = texture; }
+ImagePtr Material::GetTexture(int slot) const
+{
+	auto i = m_textures.find(slot);
+	if (i == m_textures.end())
+	{
+		return nullptr;
+	}
+	return i->second;
+}
 }  // namespace pagoda::material